1.修改trigger_cb_handle bug

+++ package/ramips/applications/bestomgw/src/dalitek/src/m1_district.c
@@ -12,7 +12,7 @@
 
 int district_create_handle(payload_t data)
 {
-	printf("district_create_handle\n");
+	fprintf(stdout,"district_create_handle\n");
 	cJSON* districtNameJson = NULL;
 	cJSON* apIdJson = NULL;
 	cJSON* apIdArrayJson = NULL;
@@ -23,6 +23,7 @@ int district_create_handle(payload_t data)
 	int row_number = 0;
 	char time[30];
 	char sql_1[200];
+    char* errorMsg = NULL;
 
 	if(data.pdu == NULL) return M1_PROTOCOL_FAILED;
 	getNowTime(time);
@@ -39,42 +40,50 @@ int district_create_handle(payload_t data)
 	char* sql = "select ID from district_table order by ID desc limit 1";
 	/*linkage_table*/
 	id = sql_id(db, sql);
+    if(sqlite3_exec(db, "BEGIN", NULL, NULL, &errorMsg)==SQLITE_OK){
+        fprintf(stdout,"BEGIN\n");
+    	/*获取收到数据包信息*/
+        districtNameJson = cJSON_GetObjectItem(data.pdu, "districtName");
+        fprintf(stdout,"districtName:%s\n",districtNameJson->valuestring);
+        apIdArrayJson = cJSON_GetObjectItem(data.pdu, "apId");
+        number1 = cJSON_GetArraySize(apIdArrayJson);
+        fprintf(stdout,"number1:%d\n",number1);
 
-	/*获取收到数据包信息*/
-    districtNameJson = cJSON_GetObjectItem(data.pdu, "districtName");
-    printf("districtName:%s\n",districtNameJson->valuestring);
-    apIdArrayJson = cJSON_GetObjectItem(data.pdu, "apId");
-    number1 = cJSON_GetArraySize(apIdArrayJson);
-    printf("number1:%d\n",number1);
+       	/*删除原有表scenario_table中的旧scenario*/
+    	sprintf(sql_1,"select ID from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);	
+    	row_number = sql_row_number(db, sql_1);
+    	fprintf(stdout,"row_number:%d\n",row_number);
+    	if(row_number > 0){
+    		sprintf(sql_1,"delete from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);				
+    		sqlite3_reset(stmt);
+    		sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
+    		while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW);
+    	}
+        
+        /*存取到数据表scenario_table中*/
+        for(i = 0; i < number1; i++){
+    		apIdJson = cJSON_GetArrayItem(apIdArrayJson, i);
+    		fprintf(stdout,"apId:%s\n",apIdJson->valuestring);
 
-   	/*删除原有表scenario_table中的旧scenario*/
-	sprintf(sql_1,"select ID from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);	
-	row_number = sql_row_number(db, sql_1);
-	printf("row_number:%d\n",row_number);
-	if(row_number > 0){
-		sprintf(sql_1,"delete from district_table where DIS_NAME = \"%s\";",districtNameJson->valuestring);				
-		sqlite3_reset(stmt);
-		sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt, NULL);
-		while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW);
-	}
-    
-    /*存取到数据表scenario_table中*/
-    for(i = 0; i < number1; i++){
-		apIdJson = cJSON_GetArrayItem(apIdArrayJson, i);
-		printf("apId:%s\n",apIdJson->valuestring);
-
-		sql = "insert into district_table(ID, DIS_NAME, AP_ID, TIME) values(?,?,?,?);";
-		printf("sql:%s\n",sql);
-		sqlite3_reset(stmt);
-		sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
-		sqlite3_bind_int(stmt, 1, id);
-		id++;
-		sqlite3_bind_text(stmt, 2, districtNameJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 3, apIdJson->valuestring, -1, NULL);
-		sqlite3_bind_text(stmt, 4, time, -1, NULL);
-		rc = thread_sqlite3_step(&stmt, db); 
-		printf("step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
-    }
+    		sql = "insert into district_table(ID, DIS_NAME, AP_ID, TIME) values(?,?,?,?);";
+    		fprintf(stdout,"sql:%s\n",sql);
+    		sqlite3_reset(stmt);
+    		sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
+    		sqlite3_bind_int(stmt, 1, id);
+    		id++;
+    		sqlite3_bind_text(stmt, 2, districtNameJson->valuestring, -1, NULL);
+    		sqlite3_bind_text(stmt, 3, apIdJson->valuestring, -1, NULL);
+    		sqlite3_bind_text(stmt, 4, time, -1, NULL);
+    		rc = thread_sqlite3_step(&stmt, db); 
+    		fprintf(stdout,"step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
+        }
+        if(sqlite3_exec(db, "COMMIT", NULL, NULL, &errorMsg) == SQLITE_OK){
+            fprintf(stdout,"END\n");
+        }
+    }else{
+        fprintf(stdout,"errorMsg:");
+    }    
+    sqlite3_free(errorMsg);
 
     sqlite3_finalize(stmt);
     sqlite3_close(db);
@@ -84,7 +93,7 @@ int district_create_handle(payload_t data)
 
 int app_req_district(int clientFd, int sn)
 {
-	printf("app_req_district\n");
+	fprintf(stdout,"app_req_district\n");
 	/*cJSON*/
     int pduType = TYPE_M1_REPORT_DISTRICT_INFO;
 
@@ -104,7 +113,7 @@ int app_req_district(int clientFd, int sn)
     pJsonRoot = cJSON_CreateObject();
     if(NULL == pJsonRoot)
     {
-        printf("pJsonRoot NULL\n");
+        fprintf(stdout,"pJsonRoot NULL\n");
         cJSON_Delete(pJsonRoot);
         return M1_PROTOCOL_FAILED;
     }
@@ -146,7 +155,7 @@ int app_req_district(int clientFd, int sn)
     /*取区域名称*/
     char* dist_name = NULL, *ap_id = NULL, *ap_name = NULL;
     sql = "select distinct DIS_NAME from district_table;";
-   	printf("sql:%s\n", sql);
+   	fprintf(stdout,"sql:%s\n", sql);
     sqlite3_reset(stmt);
     sqlite3_prepare_v2(db, sql, strlen(sql), &stmt, NULL);
     while(thread_sqlite3_step(&stmt, db) == SQLITE_ROW){
@@ -169,12 +178,12 @@ int app_req_district(int clientFd, int sn)
 	    }
 	    cJSON_AddItemToObject(devDataObject, "apInfo", apInfoArrayObject);
 	    sprintf(sql_1,"select AP_ID from district_table where DIS_NAME = \"%s\" ;",dist_name);
-	    printf("sql_1:%s\n", sql_1);
+	    fprintf(stdout,"sql_1:%s\n", sql_1);
 	    sqlite3_reset(stmt_1);
 	    sqlite3_prepare_v2(db, sql_1, strlen(sql_1), &stmt_1, NULL);
 	    while(thread_sqlite3_step(&stmt_1, db) == SQLITE_ROW){
 	    	ap_id = sqlite3_column_text(stmt_1,0);
-	    	printf("ap_id:%s\n",ap_id);
+	    	fprintf(stdout,"ap_id:%s\n",ap_id);
 		    apInfoObject = cJSON_CreateObject();
 		    if(NULL == apInfoObject)
 		    {
@@ -185,14 +194,14 @@ int app_req_district(int clientFd, int sn)
 		    cJSON_AddStringToObject(apInfoObject, "apId", ap_id);
 		    /*取出apName*/
 		    sprintf(sql_2,"select DEV_NAME from all_dev where DEV_ID = \"%s\" ;",ap_id);
-		    printf("sql_2:%s\n", sql_2);
+		    fprintf(stdout,"sql_2:%s\n", sql_2);
 		    sqlite3_reset(stmt_2);
 		    sqlite3_prepare_v2(db, sql_2, strlen(sql_2), &stmt_2, NULL);
 		    rc = thread_sqlite3_step(&stmt_2, db); 
-			printf("step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
+			fprintf(stdout,"step() return %s\n", rc == SQLITE_DONE ? "SQLITE_DONE": rc == SQLITE_ROW ? "SQLITE_ROW" : "SQLITE_ERROR"); 
 			if(rc == SQLITE_ROW){
 				ap_name = sqlite3_column_text(stmt_2,0);
-				printf("ap_name:%s\n",ap_name);
+				fprintf(stdout,"ap_name:%s\n",ap_name);
 				cJSON_AddStringToObject(apInfoObject, "apName", ap_name);
 			}
 	    }
@@ -211,7 +220,7 @@ int app_req_district(int clientFd, int sn)
         return M1_PROTOCOL_FAILED;
     }
 
-    printf("string:%s\n",p);
+    fprintf(stdout,"string:%s\n",p);
     /*response to client*/
     socketSeverSend((uint8*)p, strlen(p), clientFd);
     cJSON_Delete(pJsonRoot);
